+++ plugins/woocommerce/includes/admin/meta-boxes/views/html-product-data-attributes.php
@@ -6,16 +6,16 @@
  */
 
 use Automattic\WooCommerce\Admin\Features\Features;
-use Automattic\WooCommerce\Admin\Features\ProductVariationsClassicRedesign\Init;
+use Automattic\WooCommerce\Admin\Features\ProductVariationsClassicRedesign;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-if ( Features::exists( Init::FEATURE_ID ) ) {
+if ( Features::exists( ProductVariationsClassicRedesign::FEATURE_ID ) ) {
 	?>
 	<div id="product_attributes" class="panel wc-metaboxes-wrapper hidden">
-		<div id="<?php echo esc_attr( Init::ATTRIBUTES_ROOT_ID ); ?>"></div>
+		<div id="<?php echo esc_attr( ProductVariationsClassicRedesign::ATTRIBUTES_ROOT_ID ); ?>"></div>
 	</div>
 	<?php
 	return;
