+++ plugins/woocommerce/includes/admin/meta-boxes/views/html-product-data-variations.php
@@ -6,18 +6,18 @@
  */
 
 use Automattic\WooCommerce\Admin\Features\Features;
-use Automattic\WooCommerce\Admin\Features\ProductVariationsClassicRedesign\Init;
+use Automattic\WooCommerce\Admin\Features\ProductVariationsClassicRedesign;
 use Automattic\WooCommerce\Internal\CostOfGoodsSold\CostOfGoodsSoldController;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-if ( Features::exists( Init::FEATURE_ID ) ) {
+if ( Features::exists( ProductVariationsClassicRedesign::FEATURE_ID ) ) {
 	?>
 	<div id="variable_product_options" class="panel wc-metaboxes-wrapper hidden">
 		<div id="variable_product_options_inner">
-			<div id="<?php echo esc_attr( Init::ROOT_ID ); ?>"></div>
+			<div id="<?php echo esc_attr( ProductVariationsClassicRedesign::ROOT_ID ); ?>"></div>
 		</div>
 	</div>
 	<?php
